requirements.txt
streamlit>=1.29         #Frontend framework for web apps
supabase>=2.0.2         #Supabase client for database operations
fastapi>=0.104.1        #Backend API framework 
uvicorn>=0.24.0         #ASGI server for FastAPI
python-dotenv>=1.0.0    #Environment variable management
plotly